feat(vtkview): read and select camera view points
For vtkjs files only, if there is a "cameraViewPoints" field, the camera view points will be read
into paraview glance and saved in the store. If there are any camera view points loaded, and if the
view is a 3D view, a menu button will appear at the bottom of the VtkView, allowing the user to
select a camera view point. When a view point is selected, the view point will be set and rendered.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>