improvement(camera view points): modify source visibilities
This adds the ability to modify source visibilities for camera view points.

In a view point object, if there is a "show" array containing source names, all of those sources
will be shown when that camera view point is chosen. Additionally, if there is a "hide" array
containing source names, all of those sources will be hidden when that camera view point is chosen.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>