fix(viewpoints): allow viewpoints to be set via importBase64Dataset()
Previously, viewpoints could only be set in the store via OPEN_FILES() in store/fileLoader.js. This
is called when the user opens files via dialog in the GUI. This meant that methods like embedding a
vtk.js file in ParaView Glance html would not set the viewpoints in the store. To fix it, we made
the store accessible via the proxy manager, and we now set the viewpoints to the store via
registerReadersToProxyManager() in ReaderFactory.js.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>