fix(viewpoints): retain focal distance when switching view points
Before, when the viewpoint was changed, sometimes the focal distance would drastically change (1000
to 1, for instance). When this would happen, camera interaction including the default and first
person would result in some weird camera shaking during the interaction. Retaining the focal
distance fixes this issue, and it makes switching between camera viewpoints also appear to be much
smoother as well.