fix(bundle): ensure css and one font file get bundled
Before this fix, the css files would not be inlined, and thus, the bundle would depend on an
external css file and not work. Additionally, after the css file was inlined, no font files were
being included. Raising the url-loader limit for the font files to 300000 allows us to include just
one font file. The resulting paraview glance bundle file is ~3.7 MB, and is standalone.

fix #285