fix(movementspeed): ensure a valid movement speed is set in globalsettings
If the user attempts to view the global settings before a movement has been performed, the movement
speed will not have yet been calculated. This resulted in the movement speed being null in the
global settings. Fix this by resetting the movement speed if it is attempted to be displayed and it
is null.